# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PubWIFIprot API",
        "description": "API documentation for the PubWIFIprot backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pubwifiprot.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/networks/analyze": {
            "post": {
                "tags": [
                    "Networks"
                ],
                "summary": "Analyze list of networks from mobile device",
                "operationId": "e6e5bc38e7b71798b2cc7cddfeb90aba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "ssid": {
                                            "type": "string",
                                            "example": "CoffeeShop_WiFi"
                                        },
                                        "bssid": {
                                            "type": "string",
                                            "example": "00:11:22:33:44:55"
                                        },
                                        "signal_strength": {
                                            "type": "integer",
                                            "example": -65
                                        },
                                        "encryption_type": {
                                            "type": "string",
                                            "example": "WPA2"
                                        },
                                        "channel": {
                                            "type": "integer",
                                            "example": 6
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Analysis of networks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "ssid": {
                                                "type": "string",
                                                "example": "CoffeeShop_WiFi"
                                            },
                                            "bssid": {
                                                "type": "string",
                                                "example": "00:11:22:33:44:55"
                                            },
                                            "signal_strength": {
                                                "type": "integer",
                                                "example": -65
                                            },
                                            "encryption_type": {
                                                "type": "string",
                                                "example": "WPA2"
                                            },
                                            "channel": {
                                                "type": "integer",
                                                "example": 6
                                            },
                                            "is_trusted": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "is_suspicious": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "trust_score": {
                                                "type": "integer",
                                                "example": 85
                                            },
                                            "warnings": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/networks/check": {
            "post": {
                "tags": [
                    "Networks"
                ],
                "summary": "Check if a specific network is safe",
                "operationId": "1b7f7c4827478f860145abbd670e24ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ssid": {
                                        "type": "string",
                                        "example": "CoffeeShop_WiFi"
                                    },
                                    "bssid": {
                                        "type": "string",
                                        "example": "00:11:22:33:44:55"
                                    },
                                    "signal_strength": {
                                        "type": "integer",
                                        "example": -65
                                    },
                                    "encryption_type": {
                                        "type": "string",
                                        "example": "WPA2"
                                    },
                                    "channel": {
                                        "type": "integer",
                                        "example": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Network safety check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_safe": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "trust_score": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "warnings": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "is_trusted": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_suspicious": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}